% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPACA.R
\name{autoPACA}
\alias{autoPACA}
\title{Phenotype Aware Components Analysis with Automatic \eqn{k} Selection (autoPACA)}
\usage{
autoPACA(X, Y, pcRank = 2, thrsh = 10)
}
\arguments{
\item{X}{\eqn{n_1} by \eqn{m} matrix, where \eqn{m > n_1}; \cr
Case (foreground) input data matrix. \cr
It is recommended to normailize the feature scales as appropriate for the data modality.
E.g. quantile normalization (or other comparable approaches) for RNAseq data.}

\item{Y}{\eqn{n_0} by \eqn{m} matrix, where \eqn{m > n_0}; \cr
Control (foreground) input data matrix. \cr
It is recommended to normailize the feature scales as appropriate for the data modality.
E.g. quantile normalization (or other comparable approaches) for RNAseq data.}

\item{pcRank}{Positive integer, optional (default \eqn{2}); \cr
Number of dominant principle components to be computed for the corrected case data.}

\item{thrsh}{Positive real value, optional (default \eqn{10}); \cr
Threshold value for the maximum ratio of variance in \emph{PACA} corrected \code{X} PCs and the variance it explain in Y
which indicates the presence of residual shared variation in X.}
}
\value{
\code{autoPACA} returns a list containing the following components:
\describe{
   \item{k}{        Integer; \cr
                    the estimated number of shared axis of variation to be removed from the case data \code{X}.
   }
   \item{x}{        \eqn{n_1} by \eqn{pcRank} matrix; \cr
                    the projections / scores of the \emph{PACA} corrected case data (\code{xtil}).
   }
   \item{rotation}{  \eqn{m} by \eqn{pcRank} matrix; \cr
                     the rotation (eigenvectors)  of the \emph{PACA} corrected case data (\code{xtil}).
   }

   \item{xtil}{     \eqn{m} by \eqn{n_1} matrix; \cr
                    the \emph{PACA} corrected case data, i.e., the data with the case-specific variation only. \cr
                    NOTE: when \eqn{residOnly = TRUE}, only the \code{xtil} matrix is returned.
   }
}
}
\description{
This is an extension to the basic \code{\link{PACA}} algorithm in which the
number of shared dimensions \code{k} to be removed is automatically selected.
}
