% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eigenprcomp}
\alias{eigenprcomp}
\title{Principal Components Analysis using RcppEigen}
\usage{
eigenprcomp(x, center = TRUE, scale = FALSE, rank = NULL, tol = NULL, info = 1)
}
\arguments{
\item{x}{A numeric matrix or data frame which provides the data for the
principal components analysis.}

\item{center}{A logical value indicating whether the variables should be
shifted to be zero centered. Alternately, a vector of length
equal the number of columns of x can be supplied. The value
is passed to scale.}

\item{scale}{A logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place.
Alternatively, a vector of length equal the number of columns
of x can be supplied. The value is passed to scale.}

\item{rank}{A positive integer specifying the desired rank (i.e., number
of principal components to be computed).}

\item{tol}{A threshold for omitting principal components (see Details).}

\item{info}{Integer, optional (default: 0); \cr
Verbosity level for the log generated. \cr
0: Errors and warnings only \cr
1: Basic informational messages \cr
2: More detailed informational messages \cr
3: Debug mode, all informational log is dumped}
}
\value{
A list with class "prcomp" containing the following components:
\describe{
  \item{sdev}{The standard deviations of the principal components.}
  \item{rotation}{The matrix of variable loadings (eigenvectors).}
  \item{x}{The rotated data (the centred, scaled data multiplied by the rotation matrix).}
  \item{center}{The centering used, or FALSE.}
  \item{scale}{The scaling used, or FALSE.}
}
}
\description{
This function performs a principal components analysis on the given data matrix
using an efficient RcppEigen implementation. It returns the results as a list containing
various components of the PCA.
}
\details{
The calculation is done using singular value decomposition of the (centered
and scaled) data matrix. This is generally the preferred method for numerical
accuracy. The `tol` parameter can be used to omit components whose standard
deviations are less than or equal to `tol` times the standard deviation of
the first component.
}
