% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACA-package.R
\docType{package}
\name{PACA-package}
\alias{PACA}
\alias{PACA-package}
\title{PACA: Phenotype Aware Components Analysis}
\arguments{
\item{X}{\eqn{m} by \eqn{n_1} matrix of cases or a foregroung dataset\cr
**No missing values accepted**}

\item{Y}{\eqn{m} by \eqn{n_2} matrix of controls or a background dataset\cr
**No missing values accepted**}
}
\description{
Phenotype Aware Components Analysis (PACA) is a contrastive learning approach leveraging canonical correlation analysis to robustly capture weak sources of subphenotypic variation. PACA can be used to define de novo subtypes that are more likely to reflect molecular heterogeneity, especially in challenging cases where the phenotypic heterogeneity may be masked by a myriad of strong unrelated effects in the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adigorla/PACA}
  \item Report bugs at \url{https://github.com/adigorla/PACA/issues}
}

}
\author{
\strong{Maintainer}: Aditya Gorla \email{adityagorla@ucla.edu}

}
\keyword{internal}
