% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paca.R
\name{paca}
\alias{paca}
\alias{paca_null}
\title{Phenotype Aware Components Analysis (PACA)}
\usage{
paca(
  X,
  Y,
  k = NULL,
  scale = TRUE,
  rank = 5,
  thrsh = 10,
  ccweights = FALSE,
  info = 1
)

nullEvalPACA(X, Y, k, nPerm = 100)
}
\arguments{
\item{X}{\eqn{n_1} by \eqn{m} matrix; \cr
Case (foreground) input data matrix. \cr
It is recommended to normailize the feature scales as appropriate for the data modality.
E.g. quantile normalization (or other comparable approaches) for RNAseq data.}

\item{Y}{\eqn{n_0} by \eqn{m} matrix; \cr
Control (foreground) input data matrix. \cr
It is recommended to normailize the feature scales as appropriate for the data modality.
E.g. quantile normalization (or other comparable approaches) for RNAseq data.}

\item{k}{Positive integer, \eqn{k > 1}; \cr
Number of, \eqn{k}, dimensions of shared variation to be removed from case data \code{X}.}

\item{scale}{bool, optional (default: \eqn{TRUE}); normalize (center+scale) each matrix column-wise}

\item{rank}{Positive integer, optional (default \eqn{2}); \cr
Number of dominant principle components to be computed for the corrected case data.}

\item{thrsh}{Positive real value, optional (default \eqn{10}); \cr
Threshold value for the maximum ratio of variance in \emph{PACA} corrected \code{X} PCs and the variance it explain in Y
which indicates the presence of residual shared variation in X.}

\item{ccweights}{bool, optional (default \eqn{FALSE}); \cr
If \eqn{TRUE}, return the \emph{PACA} corrected case data (\code{xtil}) ONLY.}

\item{info}{int, optional (default: 1); verbosity for the log generated
\itemize{
\item 0 : Errors and warnings only
\item 1 : Basic Informational messages
\item 2 : More detailed Informational messages
\item 3 : Debug mode, all informational log is dumped
}}

\item{nperm}{Positive integer, optional (default \eqn{100}); \cr
Number of random permutations to build the emperical null distribution.}
}
\value{
By default, \code{paca} returns a list containing the following components:
\describe{
   \item{Xtil}{     \eqn{m} by \eqn{n_1} matrix; \cr
                    the \emph{PACA} corrected case data, i.e., the data with the case-specific variation only.
   }
   \item{U0}{       \eqn{m} by \eqn{k} matrix; \cr
                    the \emph{PACA} shared components that are removed from \eqn{X}.
   }
   \item{x}{        \eqn{n_1} by \eqn{rank} matrix; \cr
                    the projections / scores of the \emph{PACA} corrected case data (\code{Xtil}).
   }
   \item{rotation}{  \eqn{m} by \eqn{rank} matrix; \cr
                     the rotation (eigenvectors)  of the \emph{PACA} corrected case data (\code{Xtil}).
   }
}

When \eqn{ccweights = TRUE}, \code{paca} returns a list containing the CCA direction and variates along withe the \emph{PACA} principle components:
\describe{
   \item{Xtil}{     \eqn{m} by \eqn{n_1} matrix; \cr
                    the \emph{PACA} corrected case data, i.e., the data with the case-specific variation only.
   }
   \item{U0}{       \eqn{m} by \eqn{k} matrix; \cr
                    the \emph{PACA} shared components that are removed from \eqn{X}.
   }
   \item{x}{        \eqn{n_1} by \eqn{rank} matrix; \cr
                    the projections / scores of the \emph{PACA} corrected case data (\code{Xtil}).
   }
   \item{rotation}{\eqn{m} by \eqn{rank} matrix; \cr
                     the rotation (eigenvectors)  of the \emph{PACA} corrected case data (\code{Xtil}).
   }
   \item{A}{       the loadings for \eqn{X}
   }
   \item{B}{       the loadings for \eqn{Y}
   }
   \item{U}{       canonical variables of \eqn{X}, calculated by column centering \eqn{X} and projecting it on \eqn{A}
   }
   \item{V}{       canonical variables of \eqn{Y}, calculated by column centering \eqn{Y} and projecting it on \eqn{B}
   }
}

\code{nullEvalPACA} returns a list containing the following components:
\describe{
   \item{pval}{     non-negative real value; \cr
                    the significance of rejecting the null hypothesis that there is no subphenotypic structure in the case data \code{X}.
   }
   \item{empVar}{   non-negative real value; \cr
                    the variance of the top \emph{PACA} PC of the case data (\code{xtil}).
   }

   \item{nullVars}{ list of size \eqn{nPerm}; \cr
                    the variances of the top \emph{PACA} PC of each of the permuted null data.
   }
}
}
\description{
Phenotype Aware Components Analysis (PACA) is a
contrastive learning approach leveraging canonical correlation analysis to robustly capture weak sources of
subphenotypic variation. Given case-control data of any modality, PACA highlights the dominant variation in a
subspace that is not affected by background variation as a putative representation of phenotypic heterogeneity. We do so by
removing the top \code{k} components of shared variation from the cases (or foreground) \code{X}.
In the context of complex disease, PACA learns a gradient of variation unique to cases \code{X} in
a given dataset, while leveraging control samples \code{Y} for accounting for variation and imbalances of biological
and technical confounders between cases and controls.

This method applies a simple case/control label permutation approach to qunatify the
statistical significance of the presence of subphenotypic variation the cases, a givne fixed \eqn{k}.
The procedure should be able to reject the null (no subphenotypic variation) when there is sufficently
strong variation unique to the cases.
}
