% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformCCAinput}
\alias{transformCCAinput}
\title{Transform Input for CCA}
\usage{
transformCCAinput(X, Y, center = TRUE, scale = FALSE)
}
\arguments{
\item{X}{\eqn{n_1} by \eqn{m} matrix; \cr
Case (foreground) input data matrix. \cr
It is recommended to normailize the feature scales as appropriate for the data modality.
E.g. quantile normalization (or other comparable approaches) for RNAseq data.}

\item{Y}{\eqn{n_0} by \eqn{m} matrix; \cr
Control (foreground) input data matrix. \cr
It is recommended to normailize the feature scales as appropriate for the data modality.
E.g. quantile normalization (or other comparable approaches) for RNAseq data.}

\item{.center}{bool, optional (default \eqn{TRUE}); \cr
If \eqn{TRUE}, data is mean centered.}

\item{.scale}{bool, optional (default \eqn{FALSE}); \cr
If \eqn{TRUE}, data is unit scaled.}
}
\value{
\code{transformCCAinput} returns a list containing the following components:
\describe{
   \item{x}{   \eqn{m} by \eqn{n_1} matrix; \cr
               the transposed and sample centered/scaled case data \code{X}.
   }
   \item{Y}{   \eqn{m} by \eqn{n_0} matrix; \cr
               the transposed and sample centered/scaled case data \code{Y}.
   }
}
}
\description{
This method transposes the input matrices from the standard \eqn{dim(n,m)} to
the form \eqn{dim(m,n)} which PACA uses. Next, it also centers/scales the
input data along the sample axis.
}
\keyword{internal}
