% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomized-paca.R
\name{paca_r}
\alias{paca_r}
\title{Randomized Phenotype Aware Components Analysis (rPACA)}
\usage{
paca_r(X, Y, k, niter = 20, batch = 600, rank = 5)
}
\arguments{
\item{X}{\eqn{m} by \eqn{n_1} matrix; \cr
Case (foreground) input data matrix. \cr
It is recommended to normailize the feature scales as appropriate for the data modality.
E.g. quantile normalization (or other comparable approaches) for RNAseq data.}

\item{Y}{\eqn{m} by \eqn{n_0} matrix; \cr
Control (foreground) input data matrix. \cr
It is recommended to normailize the feature scales as appropriate for the data modality.
E.g. quantile normalization (or other comparable approaches) for RNAseq data.}

\item{k}{Positive integer, \eqn{k > 1}; \cr
Number of, \eqn{k}, dimensions of shared variation to be removed from case data \code{X}.}

\item{niter}{Positive integer; \cr
Number random sampling iterations.}

\item{batch}{Positive integer; \cr
Number of cases/controls to be sampled for the training set at each iteration. \cr
Note that \eqn{k < batch-1} and \eqn{batch < min{m, n1, n0}}.}

\item{rank}{Positive integer, optional (default \eqn{2}); \cr
Number of dominant principle components to be computed for the corrected case data.}
}
\value{
\code{autoPACA} returns a list containing the following components:
\describe{
   \item{x}{        \eqn{n_1} by \eqn{pcRank} matrix; \cr
                    the projections / scores of the randomized \emph{PACA} corrected case data.
   }

   \item{eigVal}{   list of size \eqn{pcRank}; \cr
                    the eigen values of the returned PC components.
   }
}
}
\description{
This is the randomized version of the basic \code{\link{PACA}} algorithm.
rPACA allows us to apply PACA in regimes where p << n, i.e., in cases where the number of
samples is greater than the number of features.
}
